+++ test/main.js
@@ -46,6 +46,7 @@ const HTML = `
       <div id="app44"></div>
       <div id="app45"></div>
       <div id="app46"></div>
+      <div id="app47"></div>
       <div id="app50"></div>
     </body>
   </html>